# Add partial AVX512 Linux support for dot product on 4-bit quantized values