@antimatter15 2023-03-20
Looks great! 

That said, I'm working on trying to minimize the number of deviations from the upstream https://github.com/ggerganov/llama.cpp repo, so there would be a more appropriate place for this PR!