@Ameobea 2023-03-20
Yeah I'll get a PR up there tomorrow as well.

In the meantime, if anyone else has access to Linux with AVX512-capable CPUs, would be great if they could test this to make sure it works on their setups as well.